Inserción de datos en algunos archivos

+++ categorias/categoria_lista.php
@@ -77,8 +77,13 @@ $resultado = $conn->query($sql);
                                     <td><?= htmlspecialchars($fila['nombre']) ?></td>
                                     <td><?= htmlspecialchars($fila['descripcion']) ?></td>
                                     <td class="text-nowrap">
-                                        <a href="categoria_editar.php?id=<?= $fila['id'] ?>" class="btn btn-sm btn-action-edit me-2">Editar</a>
-                                        <a href="categoria_eliminar.php?id=<?= $fila['id'] ?>" class="btn btn-sm btn-outline-danger fw-bold" onclick="return confirm('¿Eliminar esta categoría?');">Eliminar</a>
+                                        <?php if ($_SESSION['nivel_acceso'] == 'Administrador' || $_SESSION['nivel_acceso'] == 'Supervisor'): ?>
+                                            <a href="categoria_editar.php?id=<?= $fila['id'] ?>" class="btn btn-sm btn-action-edit me-2">Editar</a>
+                                        <?php endif; ?>
+
+                                        <?php if ($_SESSION['nivel_acceso'] == 'Administrador'): ?>
+                                            <a href="categoria_eliminar.php?id=<?= $fila['id'] ?>" class="btn btn-sm btn-outline-danger fw-bold" onclick="return confirm('¿Eliminar esta categoría?');">Eliminar</a>
+                                        <?php endif; ?>
                                     </td>
                                 </tr>
                             <?php endwhile; ?>
