+++ estudiantes/estudiante_lista.php
@@ -80,9 +80,14 @@ $resultado = $conn->query($sql);
                                     <td><?= htmlspecialchars($fila['correo']) ?></td>
                                     <td><?= htmlspecialchars($fila['carrera']) ?></td>
                                     <td><?= htmlspecialchars($fila['telefono']) ?></td>
-                                    <td class="text-nowrap">
-                                        <a href="estudiante_editar.php?id=<?= $fila['id'] ?>" class="btn btn-sm btn-action-edit me-2">Editar</a>
-                                        <a href="estudiante_eliminar.php?id=<?= $fila['id'] ?>" class="btn btn-sm btn-outline-danger fw-bold" onclick="return confirm('¿Eliminar este estudiante?');">Eliminar</a>
+                                   <td class="text-nowrap">
+                                        <?php if ($_SESSION['nivel_acceso'] == 'Administrador' || $_SESSION['nivel_acceso'] == 'Supervisor'): ?>
+                                            <a href="estudiante_editar.php?id=<?= $fila['id'] ?>" class="btn btn-sm btn-action-edit me-2">Editar</a>
+                                        <?php endif; ?>
+
+                                        <?php if ($_SESSION['nivel_acceso'] == 'Administrador'): ?>
+                                            <a href="estudiante_eliminar.php?id=<?= $fila['id'] ?>" class="btn btn-sm btn-outline-danger fw-bold" onclick="return confirm('¿Eliminar este estudiante?');">Eliminar</a>
+                                        <?php endif; ?>
                                     </td>
                                 </tr>
                             <?php endwhile; ?>
